rename bcf into bruker
