+++ hyperspy/io_plugins/bruker.py
