+++ hyperspy/tests/io/bruker_data/Hitachi_TM3030Plus.bcf
