+++ hyperspy/tests/io/bruker_data/Nope.bcf
