+++ hyperspy/tests/io/bruker_data/P45_12bit_packed_8bit.bcf
