+++ hyperspy/tests/io/bruker_data/P45_16bit.npy
