+++ hyperspy/tests/io/bruker_data/P45_the_default_job.bcf
