+++ hyperspy/tests/io/bruker_data/bcf-edx-ebsd.bcf
