+++ hyperspy/tests/io/bruker_data/bcf_v2_50x50px.bcf
