+++ hyperspy/tests/io/bruker_data/bruker_nano.spx
