+++ hyperspy/tests/io/bruker_data/extracted_from_bcf.spx
