+++ hyperspy/tests/io/bruker_data/over16bit.bcf
