+++ hyperspy/tests/io/bruker_data/test_TEM.bcf
