+++ hyperspy/tests/io/bruker_data/test_original_metadata.json
