+++ hyperspy/tests/io/test_bruker.py
@@ -29,7 +29,7 @@ def test_load_16bit():
     # test bcf from hyperspy load function level
     # some of functions can be not covered
     # it cant use cython parsing implementation, as it is not compiled
-    filename = os.path.join(my_path, 'bcf_data', test_files[0])
+    filename = os.path.join(my_path, 'bruker_data', test_files[0])
     print('testing bcf instructively packed 16bit...')
     s = load(filename)
     bse, sei, hype = s
@@ -37,20 +37,20 @@ def test_load_16bit():
     assert bse.data.dtype == np.uint16
     assert sei.data.dtype == np.uint16
     assert bse.data.shape == (75, 100)
-    np_filename = os.path.join(my_path, 'bcf_data', np_file[0])
+    np_filename = os.path.join(my_path, 'bruker_data', np_file[0])
     np.testing.assert_array_equal(hype.data[:22, :22, 222],
                                   np.load(np_filename))
     assert hype.data.shape == (75, 100, 2048)
 
 
 def test_load_16bit_reduced():
-    filename = os.path.join(my_path, 'bcf_data', test_files[0])
+    filename = os.path.join(my_path, 'bruker_data', test_files[0])
     print('testing downsampled 16bit bcf...')
     s = load(filename, downsample=4, cutoff_at_kV=10)
     bse, sei, hype = s
     # sem images never are downsampled
     assert bse.data.shape == (75, 100)
-    np_filename = os.path.join(my_path, 'bcf_data', np_file[1])
+    np_filename = os.path.join(my_path, 'bruker_data', np_file[1])
     np.testing.assert_array_equal(hype.data[:2, :2, 222],
                                   np.load(np_filename))
     assert hype.data.shape == (19, 25, 1047)
@@ -63,7 +63,7 @@ def test_load_16bit_reduced():
 
 def test_load_8bit():
     for bcffile in test_files[1:3]:
-        filename = os.path.join(my_path, 'bcf_data', bcffile)
+        filename = os.path.join(my_path, 'bruker_data', bcffile)
         print('testing simple 8bit bcf...')
         s = load(filename)
         bse, sei, hype = s
@@ -75,7 +75,7 @@ def test_load_8bit():
 
 
 def test_hyperspy_wrap():
-    filename = os.path.join(my_path, 'bcf_data', test_files[0])
+    filename = os.path.join(my_path, 'bruker_data', test_files[0])
     print('testing bcf wrap to hyperspy signal...')
     from hyperspy.exceptions import VisibleDeprecationWarning
     with pytest.warns(VisibleDeprecationWarning):
@@ -123,7 +123,7 @@ def test_hyperspy_wrap():
 
     md_ref['General']['original_filename'] = hype.metadata.General.original_filename
     filename_omd = os.path.join(my_path,
-                                'bcf_data',
+                                'bruker_data',
                                 'test_original_metadata.json')
     with open(filename_omd) as fn:
         # original_metadata:
@@ -137,7 +137,7 @@ def test_hyperspy_wrap():
 
 
 def test_hyperspy_wrap_downsampled():
-    filename = os.path.join(my_path, 'bcf_data', test_files[0])
+    filename = os.path.join(my_path, 'bruker_data', test_files[0])
     print('testing bcf wrap to hyperspy signal...')
     hype = load(filename, select_type='spectrum_image', downsample=5)
     assert_allclose(
@@ -152,50 +152,50 @@ def test_hyperspy_wrap_downsampled():
 
 
 def test_get_mode():
-    filename = os.path.join(my_path, 'bcf_data', test_files[0])
+    filename = os.path.join(my_path, 'bruker_data', test_files[0])
     s = load(filename, select_type='spectrum_image', instrument='SEM')
     assert s.metadata.Signal.signal_type == "EDS_SEM"
     assert isinstance(s, signals.EDSSEMSpectrum)
 
-    filename = os.path.join(my_path, 'bcf_data', test_files[0])
+    filename = os.path.join(my_path, 'bruker_data', test_files[0])
     s = load(filename, select_type='spectrum_image', instrument='TEM')
     assert s.metadata.Signal.signal_type == "EDS_TEM"
     assert isinstance(s, signals.EDSTEMSpectrum)
 
-    filename = os.path.join(my_path, 'bcf_data', test_files[0])
+    filename = os.path.join(my_path, 'bruker_data', test_files[0])
     s = load(filename, select_type='spectrum_image')
     assert s.metadata.Signal.signal_type == "EDS_SEM"
     assert isinstance(s, signals.EDSSEMSpectrum)
 
-    filename = os.path.join(my_path, 'bcf_data', test_files[3])
+    filename = os.path.join(my_path, 'bruker_data', test_files[3])
     s = load(filename, select_type='spectrum_image')
     assert s.metadata.Signal.signal_type == "EDS_TEM"
     assert isinstance(s, signals.EDSTEMSpectrum)
 
 
 def test_wrong_file():
-    filename = os.path.join(my_path, 'bcf_data', 'Nope.bcf')
+    filename = os.path.join(my_path, 'bruker_data', 'Nope.bcf')
     with pytest.raises(TypeError):
         load(filename)
 
 
 def test_fast_bcf():
     thingy = pytest.importorskip("hyperspy.io_plugins.unbcf_fast")
-    from hyperspy.io_plugins import bcf
+    from hyperspy.io_plugins import bruker
     for bcffile in test_files:
-        filename = os.path.join(my_path, 'bcf_data', bcffile)
-        thingy = bcf.BCF_reader(filename)
+        filename = os.path.join(my_path, 'bruker_data', bcffile)
+        thingy = bruker.BCF_reader(filename)
         for j in range(2, 5, 1):
             print('downsampling:', j)
-            bcf.fast_unbcf = True              # manually enabling fast parsing
+            bruker.fast_unbcf = True              # manually enabling fast parsing
             hmap1 = thingy.parse_hypermap(downsample=j)    # using cython
-            bcf.fast_unbcf = False            # manually disabling fast parsing
+            bruker.fast_unbcf = False            # manually disabling fast parsing
             hmap2 = thingy.parse_hypermap(downsample=j)    # py implementation
             np.testing.assert_array_equal(hmap1, hmap2)
 
 
 def test_decimal_regex():
-    from hyperspy.io_plugins.bcf import fix_dec_patterns
+    from hyperspy.io_plugins.bruker import fix_dec_patterns
     dummy_xml_positive = [b'<dummy_tag>85,658</dummy_tag>',
                           b'<dummy_tag>85,658E-8</dummy_tag>',
                           b'<dummy_tag>-85,658E-8</dummy_tag>',
@@ -210,7 +210,7 @@ def test_decimal_regex():
 
 def test__all_spx_loads():
     for spxfile in spx_files:
-        filename = os.path.join(my_path, 'bcf_data', spxfile)
+        filename = os.path.join(my_path, 'bruker_data', spxfile)
         s = load(filename)
         assert s.data.dtype == np.uint64
         assert s.metadata.Signal.signal_type == 'EDS_SEM'
